% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richlegend.R
\name{richlegend}
\alias{richlegend}
\title{Add a rich legend to a plot}
\usage{
richlegend(
  plot,
  aes_type = "fill",
  x,
  y,
  label.padding = unit(0, "lines"),
  label.size = 0,
  ...
)
}
\arguments{
\item{plot}{plot}

\item{aes_type}{"fill" or "colour"}

\item{x}{x}

\item{y}{y}

\item{label.padding}{label.padding}

\item{label.size}{label.size}

\item{...}{additional parameters passed to `ggtext::geom_richtext()`}
}
\description{
Add a rich legend to a plot
}
